* projectme: Laptop hack, fixed ParticleSystem to run on ATI Radeon Mobility HD3400, some other gcc fixed and accidential delete of data directory :-(

+++ source/projectme/editor/RenderSet.h
@@ -200,14 +200,14 @@ public:
 	
 	RenderSet* getActiveRenderSet() 
 	{ 
-		if( m_active >= 0 && m_active < m_set.size() )
+        if( m_active >= 0 && m_active < (int)m_set.size() )
 			return &m_set.at(m_active);
 		return 0; 
 	}
 
 	const RenderSet* getActiveRenderSet() const
 	{ 
-		if( m_active >= 0 && m_active < m_set.size() )
+        if( m_active >= 0 && m_active < (int)m_set.size() )
 			return &m_set.at(m_active);
 		return 0; 
 	}
